\name{coef.scoreFactory}
\alias{coef.scoreFactory}
\title{Extract coefficients of scoreFactory object}
\usage{
  coef.scoreFactory(sf, grp_by = "all", targets = "all",
    ...)
}
\arguments{
  \item{sf}{A scoreFactory object}

  \item{grp_by}{}

  \item{targets}{}
}
\description{
  Example situation:
}

