\name{densityplot.scoreFactory}
\alias{densityplot.scoreFactory}
\title{Densityplot method for scoreFactory object}
\usage{
  densityplot.scoreFactory(object, grp_by = "all",
    targets = "all", drop_scoring = TRUE, margin = FALSE,
    ...)
}
\arguments{
  \item{object}{A scoreFactory object}

  \item{grp_by}{}

  \item{targets}{}

  \item{drop_scoring}{}

  \item{margin}{}
}
\description{
  Example situation:
}

