\name{predict.scoreFactory}
\alias{predict.scoreFactory}
\title{Predict method for scoreFactory object}
\usage{
  predict.scoreFactory(object, newdata, targets = "all",
    ...)
}
\arguments{
  \item{object}{A scoreFactory object}

  \item{newdata}{}

  \item{targets}{}
}
\description{
  Example situation:
}

