\name{scoreFactory}
\alias{scoreFactory}
\title{This is a class for creating logistic regression scores using the LASSO algorithm.}
\usage{
  scoreFactory(data_arg, var_list, targets, keys = NULL,
    grp_by = "Career", pct_train = 0.8, training = NULL,
    tv_stratification, lambda = "1se", Weight = NULL,
    model_label = "scoreFactory")
}
\arguments{
  \item{data_arg}{data set}

  \item{var_list}{}

  \item{targets}{}

  \item{keys}{}

  \item{grp_by}{}

  \item{pct_train}{}

  \item{training}{}

  \item{tv_stratification}{}

  \item{lambda}{}

  \item{Weight}{}

  \item{model_label}{}
}
\description{
  Example situation:
}

