\name{summary.scoreFactory}
\alias{summary.scoreFactory}
\title{Summary of scoreFactory accuracy}
\usage{
  summary.scoreFactory(object, grp_by = "all",
    targets = "all", data_split = "all", extraction = NULL)
}
\arguments{
  \item{sf}{A scoreFactory object}

  \item{grp_by}{}

  \item{targets}{}

  \item{data_split}{}

  \item{extraction}{}
}
\description{
  Example situation:
}

